Swapped naming for sound and transcription files
